Github v0.1.0 Github

Github is an elixir wrapper around the Github 'hub' CLI tool. Right now you can use it to create repositories on Github from a local repository, and delete remote repositories.

Functions

create_repo(repo_name, opts \\ [])

Creates a repository on Github. Must be run from inside a git repository.

repo_name can be just a name or organization/name

opts include

:private :: bool() the repository will be private on Github

:description :: string() The Github project description

:remote_name :: string() Sets the name of the remote branch in .git/config

Example

Github.create_repo("github", [
private: true, 
description: "An elixir wrapper around the Github 'hub' CLI tool",
remote_name: "github"
])

delete_repo(name)

Deletes a repository on Github.

repo_name can be just a name or organization/name

Example

delete_repo("old_repo")

delete_repos(repos)

Deletes multiple repositories on Github.

each repo_name can be just a name or organization/name

Example

delete_repos(["old_repo_1", "myorg/old_repo_2"])